test(sdr): Seting up for hardware testing of USRP
We tried to do only software testing for the USRP but it was going to be really complicated and potentially not work, so we decided to write code for hardware testing instead and only run the unit tests while it is connected to the ursp. There needs to be more done, but the environment should work good for future.

Not finished at all, needs a lot more work