feat(Makefile): added seperate software testing and hardware testing to run when running tests
these changes make it so that when running tests only in software, the hardware tests are excluded. but when running tests in hardware, both software and hardware tests are run. Now there are seperate commands for running software and hardware tests.

needs to be updated when new hardware testing suites are added